Feature: Add sequence generation for Blueprints
This adds a sequence method to Blueprints, which will build a path
to the most advanced version of the combosition of the blueprint.
The steps in the path will work in a way that all requirements of
each contract are satisfied at all time and each step will only
increment the version of one component.
This is only true for non infinite components atm, components that
are added to the result of the blueprint by an infinite selector, will
just be filled up with whatever is available, so that path algorithm
only operates on the finite selectors.

This also add semver range version matching for requirements and
selectors.

Change-type: patch
Signed-off-by: Andreas Fitzek <andreas@resin.io>