Add logical operator support in templates
Handlebars supports very basic if condition checking, but it only checks for
existence of a field. There are times when we want to combine conditions in order
to generate something as part of a blueprint, without defining a completely separate
blueprint for it (like generating network config schema if a dt has a wifi chip or
a usb port to which we can connect a dongle).

Change-type: minor
Signed-off-by: Stevche Radevski <stevche@balena.io>