Use depth first search to calculate cartesian product
Using depth first search allows to return the product results as an
iterable and reduces the need to keep all results in memory.

The `cartesianProductWith` maintains the function signature to return an
array of results, but this adds a new function
`iterableCartesianProductWith` that returns the results an iterator over
the results.

Change-type: minor