Add option to return iterable to Blueprint.reproduce
This is a low memory alternative to Blueprint.reproduce which used to calculate
all results in memory before returning. Using the new depth first
cartesian product calculation, the new method can return the results as
an iterable as soon as they are available instead of keeping the full
set of combinations in memory.

Change-type: minor